% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Compute indicators based on hydroclimatic variables}
\usage{
indices(
  .data,
  group = NULL,
  daily_T_minimum = NULL,
  daily_T_maximum = NULL,
  precipitation = NULL,
  streamflow = NULL
)
}
\arguments{
\item{group}{}

\item{daily_T_minimum}{}

\item{daily_T_maximum}{}

\item{precipitation}{}

\item{streamflow}{}

\item{data}{}
}
\value{
data.frame of indices; grouped if provided
}
\description{
Compute all climatological indices on a tsibble
}
