% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mopex}
\alias{mopex}
\title{Proxy function for downloading MOPEX data from the new.noaa website}
\usage{
mopex(id = 12413500, product = "series")
}
\arguments{
\item{id}{An 8-digits station number}
}
\description{
This function downloads and prepares MOPEX datasets. 
it automatically changes -99 to NA as they are known to mark not assigned 
values.
}
