% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parde_coeff.data.frame}
\alias{parde_coeff.data.frame}
\title{Coefficients de Parde}
\usage{
parde_coeff.data.frame(.data, date, discharge)
}
\arguments{
\item{.data}{}

\item{date}{}

\item{discharge}{}
}
\description{
Coefficients de Parde
}
